defects return

+++ backend/src/main/java/org/maklashev/goldengroup/model/repositories/gypsumboard/BoardProductionRepository.java
@@ -2,12 +2,9 @@ package org.maklashev.goldengroup.model.repositories.gypsumboard;
 
 import org.maklashev.goldengroup.model.entity.production.BoardProduction;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.time.LocalDateTime;
 import java.util.List;
 
-public interface BoardProductionRepository extends JpaRepository<BoardProduction, Long> {
+public interface BoardProductionRepository extends JpaRepository<BoardProduction, Integer> {
 
 //    List<BoardProduction> findAllByGypsumBoardIdIn(List<Integer> gypsumBoardIds);
